style : main 컴포넌트 구조 수정, 레이아웃 변경, 아이콘 수정
